+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ### Changes
 * Added integration tests
 * Added **directory** argument to specify the configuration files load location
+* Added **script** argument to specify a script to start form the command line
+
 
 ### Fixes
-* Argument bug fix when using run scripts
+* Argument bug fix when using `npm run script`
 
 ## 1.16.1
 * Dev dependencies updated